refactor: 데드 코드 제거와 중복 유틸 공용화 (#225) (#230)
2026-08 위생 점검(knip + jscpd + 정성 스윕)에서 나온 소형 정리 항목 일괄 반영.

죽은 코드 제거 (프로덕션 소비처 0 확인)
- common/dto/cursor.input·pagination.input + spec 삭제 (seller/user는 자체 DTO 사용)
- user-order/wishlist/review error-messages의 DTO class-validator 이관 잔재 상수 삭제
- SDL 잔재 타입 파일 삭제: UserCartPlaceholder·ProductVisibility·ConversationSenderType
  (생성물 graphql.types.ts에는 애초 미포함 — diff 없음)
- 배럴/모듈 dead export 제거: RegionRepository·StoreRepository·NotificationPayload
  (주입하는 외부 모듈 없음 — provider 등록은 유지)

사설 복제 치환
- product-review-mappers의 사설 buildRegionLabel → 공용(@/features/store)
- auth.controller의 로컬 parseAccountId → @/global/auth 버전(공백·음수 방어가 추가된
  상위 호환 — 정상 JWT 경로에는 영향 없는 방어 강화)
- soft-delete.middleware의 로컬 isRecord → @/common/utils/type-guards

과도기 잔재 정리
- store-today-pickup.helper를 유일 소비자인 store-pickup-policy.helper로 흡수
  (spec 케이스도 policy helper spec으로 이동)
- toDateRequired 파라미터를 nullable로 넓혀 도달 불가였던 required 분기를 실질화
- DAY_MS 사설 5벌 → @/common/utils/kst-time의 export 상수로 통일
- 평점 소수 첫째 자리 반올림 5벌 → @/common/utils/rating의 roundRatingAverage 신설
  (+ 단위 spec 추가)

yarn validate 전체 green(190 suites/1659 tests), knip 잔여 dead export 0.
