+++ src/features/region/index.ts
@@ -1,3 +1,2 @@
 // cross-feature 공개 API. 단일 구현 repo라 토큰/인터페이스 없이 구체 클래스로 주입(의도적).
 export { RegionModule } from '@/features/region/region.module';
-export { RegionRepository } from '@/features/region/repositories/region.repository';
