+++ src/features/user/constants/user-order-error-messages.ts
@@ -1,5 +1,3 @@
 export const USER_ORDER_ERRORS = {
-  INVALID_LIMIT: '조회 개수는 1~50 사이여야 합니다.',
-  INVALID_OFFSET: '오프셋은 0 이상이어야 합니다.',
   ORDER_NOT_FOUND: '주문을 찾을 수 없습니다.',
 } as const;
