+++ src/features/user/constants/user-wishlist-error-messages.ts
@@ -1,5 +1,3 @@
 export const USER_WISHLIST_ERRORS = {
   PRODUCT_NOT_FOUND: '상품을 찾을 수 없습니다.',
-  INVALID_OFFSET: '오프셋은 0 이상이어야 합니다.',
-  INVALID_LIMIT: '조회 개수는 1~50 사이여야 합니다.',
 } as const;
